// src/components/Home/components/JobApplication/components/TabsApplications/components/CurriculumVitae/CvDetailView/types.ts
export interface CvDetail {
    id: string
    title: string
    slug: string
    category: string
    industry: string
    views: number
    downloads: number
    image: string
    previewImage: string
    description: string
    language: string
    usage: string
    design: string
    features: string[]
}